# advance-front/src/app/services/model/adress.ts
/**
 * Advance management
 * Advance project
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ggenier@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Adress { 
    id?: number;
    adressType: Adress.AdressTypeEnum;
    flatNumber?: string;
    number?: string;
    street: string;
    zipCode: string;
    city: string;
    digitalCode?: string;
}
export namespace Adress {
    export type AdressTypeEnum = 'PERSO' | 'PROF' | 'OTHE';
    export const AdressTypeEnum = {
        PERSO: 'PERSO' as AdressTypeEnum,
        PROF: 'PROF' as AdressTypeEnum,
        OTHE: 'OTHE' as AdressTypeEnum
    };
}