# advance-front/src/app/services/model/identification.ts
/**
 * Advance management
 * Advance project
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ggenier@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Adress } from './adress';
import { Phone } from './phone';


export interface Identification { 
    id?: number;
    adresses: Array<Adress>;
    phones: Array<Phone>;
}