# advance-front/src/app/services/model/trainingCenter.ts
/**
 * Advance management
 * Advance project
 *
 * OpenAPI spec version: 0.0.1
 * Contact: ggenier@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Identification } from './identification';


/**
 * Training center reperesentation
 */
export interface TrainingCenter { 
    id?: number;
    /**
     * Trainig identification
     */
    identification: Identification;
    libelle?: string;
}